tests/wpt/web-platform-tests/speculation-rules/prerender/cross-origin-isolated.https.html

<!DOCTYPE html>
<title>Allow crossOriginIsolated in prerendered page</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const testChannel = new BroadcastChannel('test-channel');
  t.add_cleanup(() => {
    testChannel.close();
  });
  const gotMessage = new Promise(resolve => {
    testChannel.addEventListener('message', e => resolve(e.data), {once: true});
  });

  startPrerendering('resources/cross-origin-isolated.https.html');
  assert_true(await gotMessage);
}, 'Allow crossOriginIsolated in prerendered page');

</script>
</body>
